Move optionConfig content directly into config, delete optionConfig
